Vuex renamed modules according to backend
