+++ app/store/index.js
@@ -3,11 +3,10 @@ import Vuex from 'vuex';
 
 // modules
 import cart from './modules/cart';
-import category from './modules/category';
+import catalog from './modules/catalog';
+import customer from './modules/customer';
 import navigation from './modules/navigation';
-import order from './modules/order';
 import product from './modules/product';
-import profile from './modules/profile';
 import session from './modules/session';
 import wishlist from './modules/wishlist';
 
@@ -20,11 +19,10 @@ export default new Vuex.Store({
    * */
   modules: {
     cart,
-    category,
+    catalog,
+    customer,
     navigation,
-    order,
     product,
-    profile,
     session,
     wishlist,
   },
