--- app/store/modules/category/index.js
@@ -1,10 +1,0 @@
-// store/modules/category/index.js
-
-export default {
-  namespaced: true,
-  state: {},
-  getters: {},
-  mutations: {},
-  actions: {},
-};
-
